requirements.txt
astroid==2.9.3
certifi==2021.10.8
colorama==0.4.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
platformdirs==2.5.1
toml==0.10.2
typing_extensions==4.1.1
wincertstore==0.2
wrapt==1.13.3
